Re-factoring of LocalSelendroidNode and LocalIOSNode classes.
* Extracted common methods to an abstract super class.
* Reduced visibility of LocalSelendroidNode to package visible
