+++ client/src/main/java/com/paypal/selion/platform/grid/LocalIOSNode.java
@@ -15,10 +15,6 @@
 
 package com.paypal.selion.platform.grid;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Handler;
@@ -26,14 +22,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import org.apache.commons.configuration.ConversionException;
-import org.apache.commons.io.IOUtils;
-import org.apache.http.HttpHost;
-import org.apache.http.HttpResponse;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.message.BasicHttpRequest;
-import org.json.JSONException;
-import org.json.JSONObject;
 import org.openqa.grid.common.exception.GridException;
 import org.uiautomation.ios.IOSServer;
 import org.uiautomation.ios.IOSServerConfiguration;
@@ -45,9 +33,10 @@ import com.paypal.selion.logger.SeLionLogger;
 import com.paypal.test.utilities.logging.SimpleLogger;
 
 /**
- * A singleton that is responsible for encapsulating all the logic w.r.t starting/shutting down a local ios driver node.
+ * A singleton that is responsible for encapsulating all the logic w.r.t starting/shutting down a local ios driver 
+ * node.
  */
-class LocalIOSNode implements LocalServerComponent {
+class LocalIOSNode extends AbstractNode implements LocalServerComponent {
     protected boolean isRunning = false;
     private IOSServer server = null;
     private SimpleLogger logger = SeLionLogger.getLogger();
@@ -81,7 +70,7 @@ class LocalIOSNode implements LocalServerComponent {
         try {
             int port = new LocalGridConfigFileParser().getPort() + 1;
             startIOSDriverNode(port);
-            waitForNodeToComeUp(port);
+            waitForNodeToComeUp(port,"Encountered problems when attempting to register the IOS Node to the local Grid");
             isRunning = true;
             logger.log(Level.INFO, "Attached iOSDriver node to local hub " + registrationUrl);
         } catch (Exception e) {
@@ -89,64 +78,6 @@ class LocalIOSNode implements LocalServerComponent {
             throw new GridException("Failed to start a local iOS Node", e);
         }
 
-    }
-
-    public void waitForNodeToComeUp(int port) {
-        logger.entering(port);
-        for (int i = 0; i < 5; i++) {
-            try {
-                // Sleep for 10 seconds.
-                Thread.sleep(10 * 1000);
-            } catch (InterruptedException e) {
-                throw new GridException(e.getMessage(), e);
-            }
-            if (wasNodeSpawned(port)) {
-                logger.exiting();
-                return;
-            }
-        }
-        throw new GridException("Encountered problems when attempting to register the IOS Node to the local Grid");
-    }
-
-    public boolean wasNodeSpawned(int port) {
-        logger.entering(port);
-        String endPoint = String.format("http://localhost:%d/wd/hub/status", port);
-
-        CloseableHttpClient client = HttpClientBuilder.create().build();
-        try {
-            URL url = new URL(endPoint);
-            URL api = new URL("http://" + url.getHost() + ":" + url.getPort() + "/wd/hub/status");
-            HttpHost host = new HttpHost(api.getHost(), api.getPort());
-
-            BasicHttpRequest r = new BasicHttpRequest("GET", api.toExternalForm());
-
-            HttpResponse response = client.execute(host, r);
-            if (response.getStatusLine().getStatusCode() != 200) {
-                throw new GridException("hub down or not responding. Reason : "
-                        + response.getStatusLine().getReasonPhrase());
-            }
-            JSONObject o = extractObject(response);
-            boolean status = (o.getInt("status") == 0);
-            logger.exiting(status);
-            return status;
-        } catch (Exception e) {
-            throw new GridException("Problem querying the status", e);
-        } finally {
-            IOUtils.closeQuietly(client);
-        }
-    }
-
-    private JSONObject extractObject(HttpResponse resp) throws IOException, JSONException {
-        logger.entering(resp);
-        BufferedReader rd = new BufferedReader(new InputStreamReader(resp.getEntity().getContent()));
-        StringBuilder s = new StringBuilder();
-        String line;
-        while ((line = rd.readLine()) != null) {
-            s.append(line);
-        }
-        rd.close();
-        logger.exiting(s.toString());
-        return new JSONObject(s.toString());
     }
 
     private void startIOSDriverNode(int port) throws Exception {
